requirements.txt
black>=22.1.0
eth-brownie>=1.19.0
flake8>=3.7.9
isort>=4.3.21
pre-commit>=2.4.0
tox>=3.15.1
eth_abi==2.1.1
